feat: theme Badge instead of Tag
Tag is actually a composite of badge, but our "Tag" design is currently more suitable to be a "Badge" component in chakra.
