--- frontend/src/components/Tag/index.ts
@@ -1,1 +1,0 @@
-export { Tag as default } from './Tag'
