+++ frontend/src/theme/components/index.ts
@@ -1,5 +1,6 @@
 import { RATING_THEME_KEY, RatingField } from './Field/Rating'
 import { YESNO_THEME_KEY, YesNoField } from './Field/YesNo'
+import { Badge } from './Badge'
 import { Banner } from './Banner'
 import { Button } from './Button'
 import { Form } from './Form'
@@ -11,11 +12,11 @@ import { Link } from './Link'
 import { NumberInput } from './NumberInput'
 import { Pagination, PAGINATION_THEME_KEY } from './Pagination'
 import { PhoneNumberInput } from './PhoneNumberInput'
-import { Tag } from './Tag'
 import { Textarea } from './Textarea'
 import { Toast } from './Toast'
 
 export const components = {
+  Badge,
   Banner,
   Button,
   Form,
@@ -30,6 +31,5 @@ export const components = {
   [PAGINATION_THEME_KEY]: Pagination,
   [RATING_THEME_KEY]: RatingField,
   [YESNO_THEME_KEY]: YesNoField,
-  Tag,
   Toast,
 }
